feat(schedule) restructured and scheule is finished
The code was restructured and the schedule is finished.
Now the files group similar functions functions and the schedule system is finished.
the schedule system create a new threat that once per day check the schedules at a pre-defined time,
this time can be seted in the variable "schedule_verification_time" in config.py.
The schedule system is not yet implemented and not tested.
to implemente you have to import  schedule in app.py  and run start_schedule().
