+++ asterisk/config.py
@@ -1,0 +1,4 @@
+URL = 'http://127.0.0.1:5000'
+HEADERS = {'Content-Type': 'application/json'}
+ADM_USER = 'admin'
+ADM_PSSW = 'adm123'
